@@ -10,57 +10,54 @@ environment:
   sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
-  adapty_flutter: ^2.4.4
-  animate_do: ^3.0.2
-  bloc: ^8.1.2
-  carrier_info: 2.0.6
-  cupertino_icons: ^1.0.5
-  cupertino_onboarding: ^1.1.1
-  dart_ping: ^8.0.1
-  dart_ping_ios: ^3.0.0
+  adapty_flutter: ^2.9.2
+  animate_do: ^3.3.3
+  bloc: ^8.1.3
+  carrier_info: ^2.0.8
+  cupertino_icons: ^1.0.6
+  cupertino_onboarding: ^1.2.0
+  dart_ping: ^9.0.1
+  dart_ping_ios: ^4.0.2
   dynamic_color: ^1.4.0
   enum_to_string: ^2.0.1
   equatable: ^2.0.5
-  flex_color_scheme: ^7.2.0
+  flex_color_scheme: ^7.3.1
   flutter:
     sdk: flutter
-  flutter_bloc: ^8.1.3
-  flutter_platform_widgets: ^3.3.5
-  flutter_svg: ^2.0.7
-  font_awesome_flutter: ^10.5.0
-  google_fonts: ^5.1.0
-  google_maps_flutter: ^2.4.0
-  google_mobile_ads: ^3.0.0
+  flutter_bloc: ^8.1.4
+  flutter_platform_widgets: ^6.1.0
+  flutter_svg: ^2.0.10+1
+  font_awesome_flutter: ^10.7.0
+  google_fonts: ^6.1.0
+  google_maps_flutter: ^2.5.3
+  google_mobile_ads: ^4.0.0
   hive: ^2.2.3
   hive_flutter: ^1.1.0
-  http: ^1.0.0
-  hydrated_bloc: ^9.1.2
-  in_app_review: ^2.0.6
-  introduction_screen: ^3.1.11
+  http: ^1.2.0
+  hydrated_bloc: ^9.1.4
+  in_app_review: ^2.0.8
+  introduction_screen: ^3.1.12
   json_annotation: ^4.8.1
   json_serializable: ^6.7.1
-  # lan_scanner: ^3.5.0
-  meta: ^1.8.0
-  mockingjay: ^0.3.0
-  mocktail: ^0.3.0
+  lan_scanner: ^4.0.0+1
+  mocktail: ^1.0.3
   modal_bottom_sheet: ^3.0.0-pre
-  network_info_plus: ^4.0.1
-  network_tools: ^3.2.1
-  package_info_plus: ^4.0.2
-  path_provider: ^2.0.15
-  permission_handler: ^10.4.3
-  sentry_flutter: ^7.9.0
-  url_launcher: ^6.1.12
-  wake_on_lan: ^3.2.0+1
-  wiredash: ^1.7.2
+  network_info_plus: ^4.1.0
+  package_info_plus: ^5.0.1
+  path_provider: ^2.1.2
+  permission_handler: ^11.3.0
+  sentry_flutter: ^7.16.1
+  url_launcher: ^6.2.5
+  wake_on_lan: ^4.1.1
+  wiredash: ^2.0.0
 
 dev_dependencies:
-  bloc_test: ^9.1.3
-  build_runner: ^2.4.6
+  bloc_test: ^9.1.6
+  build_runner: ^2.2.0
   flutter_test:
     sdk: flutter
   import_sorter: ^4.6.0
-  very_good_analysis: ^5.0.0+1
+  very_good_analysis: ^5.1.0
 
 flutter:
   uses-material-design: true
