@@ -1,33 +1,24 @@
 // Dart imports:
 import 'dart:async';
-import 'dart:isolate';
+import 'dart:io';
 
 // Package imports:
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 import 'package:network_info_plus/network_info_plus.dart';
-import 'package:network_tools/network_tools.dart';
-
-// Project imports:
-import 'package:network_arch/lan_scanner/repository/lan_scanner_repository.dart';
+import 'package:lan_scanner/lan_scanner.dart';
 
 part 'lan_scanner_event.dart';
 part 'lan_scanner_state.dart';
 
 class LanScannerBloc extends Bloc<LanScannerEvent, LanScannerState> {
-  LanScannerBloc(this._lanScannerRepository)
-      : super(const LanScannerInitial()) {
+  LanScannerBloc() : super(const LanScannerInitial()) {
     on<LanScannerStarted>(_onStarted);
     on<LanScannerProgressUpdated>(_onProgressUpdated);
     on<LanScannerHostFound>(_onHostFound);
     on<LanScannerStopped>(_onStopped);
     on<LanScannerCompleted>(_onCompleted);
   }
-
-  final LanScannerRepository _lanScannerRepository;
-
-  Isolate? _scanIsolate;
 
   Future<void> _onStarted(
     LanScannerStarted event,
@@ -36,29 +27,15 @@ class LanScannerBloc extends Bloc<LanScannerEvent, LanScannerState> {
     final ip = await NetworkInfo().getWifiIP();
     final subnet = ip!.substring(0, ip.lastIndexOf('.'));
 
-    emit(const LanScannerRunStart());
+    final scanner = LanScanner();
 
-    final receivePort = ReceivePort();
-    _scanIsolate = await Isolate.spawn(
-      _lanScannerRepository.startScanning,
-      [receivePort.sendPort, subnet],
-    );
-
-    await for (final message in receivePort) {
-      if (message is ActiveHost) {
-        add(LanScannerHostFound(message));
-      }
-
-      if (message is double) {
-        add(LanScannerProgressUpdated(message));
-
-        if (message == 100.0) {
-          add(LanScannerCompleted());
-        }
-      }
+    if (Platform.isIOS) {
+      await scanner.quickIcmpScanSync(subnet);
+    } else {
+      await scanner.quickIcmpScanAsync(subnet);
     }
 
-    Isolate.exit();
+    emit(const LanScannerRunStart());
   }
 
   void _onProgressUpdated(
@@ -76,8 +53,6 @@ class LanScannerBloc extends Bloc<LanScannerEvent, LanScannerState> {
   }
 
   void _onStopped(LanScannerStopped event, Emitter<LanScannerState> emit) {
-    _scanIsolate?.kill();
-
     emit(const LanScannerRunStop());
   }
 
